feat(checkout): CHECKOUT-6340 new all-in-one sample test
