--- tests/polly.global.setup.ts
@@ -1,137 +1,0 @@
-import { Page } from '@playwright/test';
-import { MODE, Polly } from '@pollyjs/core';
-import FSPersister from '@pollyjs/persister-fs';
-import path from 'path';
-import { PlaywrightAdapter } from 'polly-adapter-playwright';
-
-import { checkout, checkoutConfig, formFields, order } from './api.mock';
-
-interface PollyOptions {
-    playwrightContext: Page;
-    recordingName: string;
-    storeURL?: string;
-}
-
-async function pollyInitializer(option: PollyOptions): Promise<Polly> {
-    const { playwrightContext: page, recordingName, storeURL } = option;
-    const mode: MODE = process.env.MODE?.toLowerCase() === 'record' ? 'record' : 'replay';
-
-    Polly.register(PlaywrightAdapter);
-    Polly.register(FSPersister);
-
-    const polly = new Polly(recordingName);
-
-    polly.configure({
-        mode,
-        // logLevel: 'info',
-        recordIfMissing: true,
-        flushRequestsOnStop: true,
-        recordFailedRequests: true,
-        adapters: ['playwright'],
-        adapterOptions: {
-            playwright: {
-                context: page,
-            },
-        },
-        persister: 'fs',
-        persisterOptions: {
-            keepUnusedRequests: false,
-            disableSortingHarEntries: true,
-            fs: {
-                recordingsDir: path.join(__dirname, './_har/'),
-            },
-        },
-        matchRequestsBy: {
-            headers: false,
-            method: true,
-            body: true,
-            order: false,
-            url: true,
-        },
-    });
-
-    if (mode === 'record') {
-        // hack for BC dev store's root certificate issue during recording HAR
-        // https://stackoverflow.com/questions/31673587/error-unable-to-verify-the-first-certificate-in-nodejs
-        process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0';
-
-        polly.server.any().on('beforePersist', (_, recording) => {
-            // recording.request.headers = recording.request.headers.map((header: { name: string; value: string }) => {
-            //     switch (header.name) {
-            //         case 'authorization':
-            //             return {...header, value: dummyData};
-            //
-            //         case 'cookie':
-            //             return {...header, value: dummyData};
-            //
-            //         case 'x-xsrf-token':
-            //             return {...header, value: dummyData};
-            //
-            //         default:
-            //             return header;
-            //     }
-            // });
-            // recording.response.headers = recording.response.headers.map((header: { name: string; value: string }) => {
-            //     switch (header.name) {
-            //         case 'set-cookie':
-            //             return {...header, value: dummyData};
-            //
-            //         case 'token':
-            //             return {...header, value: dummyData};
-            //
-            //         case 'x-session-hash':
-            //             return {...header, value: dummyData};
-            //
-            //         default:
-            //             return header;
-            //     }
-            // });
-
-            const removeSensitiveHeaders = (headers: []): {} => {
-                const dummyData = '*';
-                const sensitiveHeaderNames = [
-                    'authorization',
-                    'cookie',
-                    'set-cookie',
-                    'token',
-                    'x-session-hash',
-                    'x-xsrf-token',
-                ];
-
-                return headers.map((header: { name: string; value: string }) => {
-                    if (sensitiveHeaderNames.includes(header.name)) {
-                        return { ...header, value: dummyData };
-                    }
-
-                    return header;
-                });
-            };
-
-            recording.request.headers = removeSensitiveHeaders(recording.request.headers);
-            recording.response.headers = removeSensitiveHeaders(recording.response.headers);
-        });
-    }
-
-    if (mode === 'replay' && storeURL) {
-        // Serving static files
-        await page.route('/', route => route.fulfill( {status: 200, path: './tests/_support/index.html' } ));
-        await page.route('**/order-confirmation', route => route.fulfill( {status: 200, path: './tests/_support/orderConfirmation.html' } ));
-        await page.route('**/ablebrewingsystem4.1647253530.190.285.jpg?c=1', route => route.fulfill( {status: 200, path: './tests/_support/product.jpg' } ));
-
-        // API mockups managed by checkout team
-        await page.route('**/api/storefront/checkout/**', route => route.fulfill( {status: 200, contentType: 'application/json', body: JSON.stringify(checkout) } ));
-        await page.route('**/api/storefront/form-fields', route => route.fulfill( {status: 200, contentType: 'application/json', body: JSON.stringify(formFields) } ));
-        await page.route('**/api/storefront/checkout-settings', route => route.fulfill( {status: 200, contentType: 'application/json', body: JSON.stringify(checkoutConfig) } ));
-        await page.route('**/api/storefront/orders/**', route => route.fulfill( {status: 200, contentType: 'application/json', body: JSON.stringify(order) } ));
-
-        // Replace URLs to match URLs in recording
-        polly.server.any().on('request', req => {
-            req.url = req.url.replace('http://localhost:' + process.env.PORT, storeURL);
-            console.log('😃 REPLAY:: ' + req.url);
-        });
-    }
-
-    return polly;
-}
-
-export { pollyInitializer };
